LayoutTests/accessibility/ios-simulator/table-ancestory-expected.txt

c
a
b
1
This tests that objects in various kinds of tables report they are in the table.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button1.isInTable is true
PASS button2.isInTable is true
PASS button3.isInTable is false
PASS successfullyParsed is true

TEST COMPLETE
